hw4/requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-google-genai>=0.0.6
openai-whisper>=20231117
gtts>=2.4.0
python-dotenv>=1.0.0

# Audio processing
soundfile>=0.12.1
librosa>=0.10.1

# Web interface (optional - for file upload)
gradio>=4.7.1

# Utilities
numpy>=1.24.3
matplotlib>=3.7.2
pygame>=2.5.0
